test(pricing,cache): move matching/cache/global-cache tests into src
Merge test_pricing_matching.rs into the existing `mod tests` in
src/pricing/matching.rs, append test_pricing_cache.rs as a separate
`mod serialization_tests` inside src/pricing/cache.rs (leaving the
existing `mod parser_tests` untouched), and move test_cache_global.rs
into a new `mod tests` inside src/cache/mod.rs.

Keeps the two logical groups under `pricing/cache.rs` separate
(parser vs serialization) so each block stays focused on one concern.
